move class to new models class file move utiltity to new utilities class file
